I have extracted `ClosureMockURLProtocol` to the Mocks directory to avoid duplicate definitions.
🎯 What: I extracted `ClosureMockURLProtocol` into a dedicated file (`OpenConeTests/Mocks/ClosureMockURLProtocol.swift`) and removed its duplicate definitions from `CredentialValidatorTests.swift` and `PineconeServiceHealthCheckTests.swift`.
💡 Why: In my previous commit, I only renamed the class but didn't remove the duplicate class definitions themselves across different test files, which was still causing the "invalid redeclaration" error during compilation in CI. By moving the mock definition to a single file, I ensured it's only declared once and reused across the testing targets.
✅ Verification: I searched the codebase to confirm that `ClosureMockURLProtocol` is only declared once in the entire test suite. I also checked that references remain valid and properly formatted in tests.
✨ Result: The compilation should now complete successfully without any duplicate definition errors in the test targets.
