+++ OpenConeTests/Services/PineconeServiceHealthCheckTests.swift
@@ -1,35 +1,6 @@
 import XCTest
 @testable import OpenCone
 
-final class ClosureMockURLProtocol: URLProtocol {
-    static var requestHandler: ((URLRequest) throws -> (HTTPURLResponse, Data))?
-
-    override class func canInit(with request: URLRequest) -> Bool {
-        return true
-    }
-
-    override class func canonicalRequest(for request: URLRequest) -> URLRequest {
-        return request
-    }
-
-    override func startLoading() {
-        guard let handler = ClosureMockURLProtocol.requestHandler else {
-            fatalError("Handler is unavailable.")
-        }
-
-        do {
-            let (response, data) = try handler(request)
-            client?.urlProtocol(self, didReceive: response, cacheStoragePolicy: .notAllowed)
-            client?.urlProtocol(self, didLoad: data)
-            client?.urlProtocolDidFinishLoading(self)
-        } catch {
-            client?.urlProtocol(self, didFailWithError: error)
-        }
-    }
-
-    override func stopLoading() {
-    }
-}
 
 @MainActor
 final class PineconeServiceHealthCheckTests: XCTestCase {
